+++ build.rs
@@ -1,5 +1,3 @@
-extern crate gl_generator;
-
 use gl_generator::{Api, Fallbacks, Profile, Registry};
 use std::env;
 use std::fs::File;