+++ src/os/unix.rs
@@ -6,9 +6,11 @@
     target_os = "openbsd"
 ))]
 
-pub use api::egl::ffi::EGLContext;
-pub use api::glx::ffi::GLXContext;
-pub use platform::RawHandle;
+pub use crate::api::egl::ffi::EGLContext;
+pub use crate::api::glx::ffi::GLXContext;
+use crate::os::ContextTraitExt;
+pub use crate::platform::RawHandle;
+use crate::Context;
 
 pub use winit::os::unix::EventsLoopExt;
 pub use winit::os::unix::MonitorIdExt;
@@ -16,9 +18,6 @@ pub use winit::os::unix::WindowBuilderExt;
 pub use winit::os::unix::WindowExt;
 pub use winit::os::unix::XNotSupported;
 pub use winit::os::unix::XWindowType;
-
-use os::ContextTraitExt;
-use Context;
 
 use std::os::raw;