+++ src/os/windows.rs
@@ -1,17 +1,16 @@
 #![cfg(target_os = "windows")]
 
+pub use crate::api::egl::ffi::EGLContext;
+use crate::os::ContextTraitExt;
+pub use crate::platform::RawHandle;
+use crate::Context;
+
 pub use winapi::shared::windef::HGLRC;
 pub use winit::os::windows::{
     DeviceIdExt, MonitorIdExt, WindowBuilderExt, WindowExt,
 };
 
-pub use api::egl::ffi::EGLContext;
-pub use platform::RawHandle;
-
 use std::os::raw;
-
-use os::ContextTraitExt;
-use Context;
 
 impl ContextTraitExt for Context {
     type Handle = RawHandle;