+++ src/platform/android/mod.rs
@@ -1,5 +1,4 @@
 #![cfg(target_os = "android")]
 
+pub use crate::api::android::*;
 pub use winit::EventsLoop;
-
-pub use api::android::*;