+++ src/platform/ios/mod.rs
@@ -1,3 +1,3 @@
 #![cfg(target_os = "ios")]
 
-pub use api::ios::*;
+pub use crate::api::ios::*;